:recycle: [Refactor]: 헤더에서 로그인 버튼 컴포넌트 파일 분리
